OMS7-N / Substratum support: updated to 2017-01-13
Theme Engine Port

Assets such as composed icons and ringtones need to be accessed
by apps.  This patch adds the policy needed to facilitate this.

Signed-off-by: d34d <clark@cyngn.com>

Theme Engine Port

Assets such as composed icons and ringtones need to be accessed
by apps.  This patch adds the policy needed to facilitate this.

Signed-off-by: d34d <clark@cyngn.com>